requirements.txt
yacs==0.1.8
numpy-quaternion==2021.8.30.10.33.11

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu102
torchvision==0.10.0+cu102
tensorboardX==2.2

scipy==1.7.1
debugpy==1.4.1
debugpy-run==1.2
opencv-python==4.5.3.56
matplotlib==3.4.2
tqdm==4.62.2
scikit-image==0.18.3
open3d==0.9.0.0
tqdm==4.62.2
einops==0.3.2
torch-scatter==1.4.0
kornia==0.6.1

#-f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu102_pyt190/download.html
#pytorch3d==0.5.0

